Refactor: 명시적 생성자 제거 및 Lombok 어노테이션 적용
- 명시적으로 작성된 생성자를 제거하고
  Lombok의 @RequiredArgsConstructor 어노테이션으로 대체
- 코드 간결성 향상 및 불필요한 보일러플레이트 제거
