+++ src/main/java/com/itcen/censquare/domain/auth/oauth/CustomUserDetails.java
@@ -4,18 +4,16 @@ import com.itcen.censquare.domain.member.entity.Member;
 import java.util.Collection;
 import java.util.Collections;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 @Getter
+@RequiredArgsConstructor
 public class CustomUserDetails implements UserDetails {
 
   private final Member member;
-
-  public CustomUserDetails(Member member) {
-    this.member = member;
-  }
 
   @Override
   public Collection<? extends GrantedAuthority> getAuthorities() {
