--- src/main/java/com/itcen/censquare/domain/member/entity/Role.java
@@ -1,16 +1,0 @@
-package com.itcen.censquare.domain.member.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum Role {
-  USER("ROLE_USER"),
-  ADMIN("ROLE_ADMIN");
-
-  private final String authority;
-
-  Role(String authority) {
-    this.authority = authority;
-  }
-
-}
