+++ src/main/java/com/itcen/censquare/global/config/SecurityConfig.java
@@ -5,6 +5,7 @@ import com.itcen.censquare.domain.auth.jwt.JwtProvider;
 import com.itcen.censquare.domain.auth.oauth.CustomOAuth2UserService;
 import com.itcen.censquare.domain.auth.oauth.OAuth2SuccessHandler;
 import com.itcen.censquare.domain.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -14,22 +15,14 @@ import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
 @Configuration
+@RequiredArgsConstructor
 @EnableWebSecurity
 public class SecurityConfig {
 
   private final JwtProvider jwtProvider;
   private final MemberRepository memberRepository;
-
   private final CustomOAuth2UserService customOAuth2UserService;
   private final OAuth2SuccessHandler oAuth2SuccessHandler;
-
-  public SecurityConfig(JwtProvider jwtProvider, MemberRepository memberRepository,
-      CustomOAuth2UserService customOAuth2UserService, OAuth2SuccessHandler oAuth2SuccessHandler) {
-    this.jwtProvider = jwtProvider;
-    this.memberRepository = memberRepository;
-    this.customOAuth2UserService = customOAuth2UserService;
-    this.oAuth2SuccessHandler = oAuth2SuccessHandler;
-  }
 
   @Bean
   public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
